review closure

+++ pallets/grants/src/migrations.rs
@@ -132,7 +132,7 @@ pub mod v1 {
 
 				stored_data
 					.iter()
-					.for_each(|(_account, old_vesting)| assert!(old_vesting.len() as u32 <= T::MaxSchedule::get()));
+					.for_each(|(_account, old_vesting)| assert!(old_vesting.len() <= T::MaxSchedule::get() as usize));
 
 				log::info!(
 					"pre_upgrade[{:#?}]=> VestingSchedules map count :: [{:#?}]",
